update Scrims.jsx, ScrimDetail.jsx, remove bootcampImg.js, move scrim components to /scrim_components
