+++ client/src/utils/getSortedScrims.js
@@ -1,11 +1,13 @@
 //  scrims sorting
 
+// descending
 export const showLatestFirst = (scrims) =>
   scrims.sort(
     (a, b) =>
       new Date(b.gameStartTime).getTime() - new Date(a.gameStartTime).getTime()
   );
 
+// ascending
 export const showEarliestFirst = (scrims) =>
   scrims.sort(
     (a, b) =>
